chore: improve according to PR comments
