+++ autonomi/src/client/mod.rs
@@ -7,14 +7,13 @@
 // permissions and limitations relating to use of the SAFE Network Software.
 
 pub mod address;
-pub mod error;
 
 #[cfg(feature = "data")]
 pub mod archive;
 #[cfg(feature = "data")]
 pub mod data;
 #[cfg(feature = "fs")]
-pub mod files;
+pub mod fs;
 #[cfg(feature = "registers")]
 pub mod registers;
 #[cfg(feature = "vault")]
