+++ autonomi/tests/fs.rs
@@ -21,7 +21,6 @@ use walkdir::WalkDir;
 
 // With a local evm network, and local network, run:
 // EVM_NETWORK=local cargo test --features="fs,local" --package autonomi --test file
-#[cfg(feature = "fs")]
 #[tokio::test]
 async fn dir_upload_download() -> Result<()> {
     let _log_appender_guard =
